requirements.txt
openai~=0.28.0
python-dotenv~=1.0.0
colorama~=0.4.6
flask~=2.3.0
flask-socketio~=5.3.0
gevent~=23.9.0
